refactor: CardAddComplete 페이지 컴포넌트명을 CardEditNickname으로 변경
- CardEditNickname 페이지에서 Spinner가 표시될 때 취소 버튼도 같이 보이지 않도록 변경

- CardEditNickname 페이지에서 닉네임을 수정하는 상황일 때의메시지를 출력하도록 수정
