+++ src/constants/message.js
@@ -6,6 +6,7 @@ const MESSAGE = {
   CARD_ADD_FAILED: '카드를 추가하지 못했어요! 다시 시도해주세요',
   ERROR_PAGE_DEFAULT: '문제가 있는 것 같아요!\n잠시 후에 다시 시도해주세요!',
   ERROR_PAGE_CARD_LIST: '카드 목록을 불러오는 데 문제가 생겼어요!\n잠시 후에 다시 시도해주세요!',
+  NICKNAME_EDIT_FAILED: '닉네임을 변경하는 데 문제가 생겼어요! 잠시 후에 다시 시도해주세요!',
   NICKNAME_SUBMIT_FAILED:
     '카드는 등록되었지만 닉네임을 설정하지 못했어요! 추후에 닉네임을 수정해주세요',
   CARD_NUMBER_NOT_ENTERED_HALF: '카드번호 앞 8자리를 먼저 입력해주세요',
